requirements.txt
torch>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pydantic>=2.0.0
joblib>=1.3.0
azure-ai-ml>=1.15.0
azure-identity>=1.15.0
